Security added

+++ app/Http/Controllers/MeetingSecretaryController.php
@@ -20,21 +20,48 @@ class MeetingSecretaryController extends Controller
     {
         $instance = \Instantiation::instance();
 
-        $meetings = Auth::user()->secretary->comittee->meetings()->get();
+            if(auth('api')->user()->secretary) {
+                $meetings = Auth::user()->secretary->comittee->meetings()->get();
+                $comittee_meeting = $meetings->comittee;
 
-        return view('meeting.list',
-            ['instance' => $instance, 'meetings' => $meetings]);
+                $secretary = auth('api')->user()->secretary;
+                $comittee_secretary = $secretary->comittee;
+
+                if($comittee_meeting == $comittee_secretary) {
+                    return view('meeting.list',
+                        ['instance' => $instance, 'meetings' => $meetings]);
+                } else {
+                    return response()->json([
+                        'success' => false,
+                        'message' => 'El usuario no tiene permisos sobre este comité.'
+                    ], 403);
+                }
+            } else {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'El usuario no es secretario.'
+                ], 403);
+            }
     }
 
     public function create()
     {
         $instance = \Instantiation::instance();
 
-        $users = User::orderBy('surname')->get();
-        $defaultlists = Auth::user()->secretary->default_lists;
+        if(auth('api')->user()->secretary) {
+            $users = User::orderBy('surname')->get();
+            $defaultlists = Auth::user()->secretary->default_lists;
 
-        return view('meeting.createandedit',
-            ['instance' => $instance, 'users' => $users, 'defaultlists' => $defaultlists, 'route' => route('secretary.meeting.new',$instance)]);
+            return view('meeting.createandedit',
+                ['instance' => $instance, 'users' => $users, 'defaultlists' => $defaultlists, 'route' => route('secretary.meeting.new',$instance)]);
+
+
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'El usuario no es secretario.'
+            ], 403);
+        }
     }
 
     public function new(Request $request)
@@ -61,32 +88,52 @@ class MeetingSecretaryController extends Controller
         ]);
 
         $meeting->comittee()->associate(Auth::user()->secretary->comittee);
+        $userid = $meeting->comittee->secretary->user->id;
 
-        $meeting->save();
+        if(auth('api')->user()->secretary){
+            $meeting->save();
 
-        // Asociamos los usuarios a la reunión
-        $users_ids = $request->input('users',[]);
+            // Asociamos los usuarios a la reunión
+            $users_ids = $request->input('users',[]);
 
-        foreach($users_ids as $user_id)
-        {
+            foreach($users_ids as $user_id)
+            {
 
-            $user = User::find($user_id);
-            $meeting->users()->attach($user);
+                $user = User::find($user_id);
+                $meeting->users()->attach($user);
 
+            }
+            return redirect()->route('secretary.meeting.list',$instance)->with('success', 'Reunión creada con éxito.');
+        } else {
+            return response()->json('El usuario no es secretario');
         }
-
-        return redirect()->route('secretary.meeting.list',$instance)->with('success', 'Reunión creada con éxito.');
-
     }
 
     public function edit($instance,$id)
     {
-        $meeting = Meeting::find($id);
-        $users = User::orderBy('surname')->get();
-        $defaultlists = Auth::user()->secretary->default_lists;
 
-        return view('meeting.createandedit',
-            ['instance' => $instance, 'meeting' => $meeting, 'edit' => true, 'users' => $users, 'defaultlists' => $defaultlists, 'route' => route('secretary.meeting.save',$instance)]);
+        if(auth('api')->user()->secretary) {
+            $meeting = Meeting::find($id);
+            $comittee_meeting = $meeting->comittee;
+
+            $users = User::orderBy('surname')->get();
+            $defaultlists = Auth::user()->secretary->default_lists;
+
+            $secretary = auth('api')->user()->secretary;
+            $comittee_secretary = $secretary->comittee;
+
+            if($comittee_meeting == $comittee_secretary) {
+                return view('meeting.createandedit',
+                    ['instance' => $instance, 'meeting' => $meeting, 'edit' => true, 'users' => $users, 'defaultlists' => $defaultlists, 'route' => route('secretary.meeting.save',$instance)]);
+            } else {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'El usuario no tiene permisos sobre este comité.'
+                ], 403);
+            }
+        } else {
+            return response()->json('El usuario no es secretario.');
+        }
     }
 
     public function defaultlist($instance,$id)
@@ -116,35 +163,67 @@ class MeetingSecretaryController extends Controller
         $meeting->place = $request->input('place');
         $meeting->datetime = $request->input('date')." ".$request->input('time');
 
-        $meeting->save();
 
-        // Asociamos los usuarios a la reunión
-        $users_ids = $request->input('users',[]);
 
-        // eliminamos usuarios antiguos de la reunión
-        foreach($meeting->users as $user)
-        {
-            $meeting->users()->detach($user);
+        if(auth('api')->user()->secretary){
+            $comittee_meeting = $meeting->comittee;
+
+            $secretary = auth('api')->user()->secretary;
+            $comittee_secretary = $secretary->comittee;
+
+            if($comittee_meeting == $comittee_secretary) {
+                $meeting->save();
+
+                // Asociamos los usuarios a la reunión
+                $users_ids = $request->input('users',[]);
+
+                // eliminamos usuarios antiguos de la reunión
+                foreach($meeting->users as $user)
+                {
+                    $meeting->users()->detach($user);
+                }
+
+                // agregamos los usuarios nuevos de la reunión
+                foreach($users_ids as $user_id)
+                {
+                    $user = User::find($user_id);
+                    $meeting->users()->attach($user);
+                }
+
+                return redirect()->route('secretary.meeting.list',$instance)->with('success', 'Reunión editada con éxito.');
+            } else {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'El usuario no tiene permisos sobre este comité.'
+                ], 403);
+            }
+        } else {
+            return response()->json('El usuario no es secretario.');
         }
-
-        // agregamos los usuarios nuevos de la reunión
-        foreach($users_ids as $user_id)
-        {
-            $user = User::find($user_id);
-            $meeting->users()->attach($user);
-        }
-
-        return redirect()->route('secretary.meeting.list',$instance)->with('success', 'Reunión editada con éxito.');
-
     }
 
     public function remove(Request $request)
     {
-        $meeting = Meeting::find($request->_id);
         $instance = \Instantiation::instance();
 
-        $meeting->delete();
+        if(auth('api')->user()->secretary) {
+            $meeting = Meeting::find($request->_id);
+            $comittee_meeting = $meeting->comittee;
 
-        return redirect()->route('secretary.meeting.list',$instance)->with('success', 'Reunión eliminada con éxito.');
+            $secretary = auth('api')->user()->secretary;
+            $comittee_secretary = $secretary->comittee;
+
+            if($comittee_meeting == $comittee_secretary) {
+                $meeting->delete();
+                return redirect()->route('secretary.meeting.list',$instance)->with('success', 'Reunión eliminada con éxito.');
+            } else {
+                return response()->json([
+                    'success' => false,
+                    'message' => 'El usuario no tiene permisos sobre este comité.'
+                ], 403);
+            }
+        } else {
+            return response()->json('El usuario no es secretario.');
+        }
     }
 }
